refactor: ReactNodeList를 반환하는 함수 컴포넌트 내부에 작성
